+++ include/Function/Function.h
@@ -13,6 +13,6 @@
 #include "FXNValue.h"
 #include "FXNValueMap.h"
 #include "FXNConfiguration.h"
-#include "FXNProfile.h"
+#include "FXNPrediction.h"
 #include "FXNPredictor.h"
 #include "FXNVersion.h"